refactor: move modal code to separate components
