+++ packages/api-client/src/components/ImportCollection/DownloadLink.vue
@@ -1,0 +1,10 @@
+<template>
+  <div class="text-sm mt-4">
+    Don’t have the Scalar app?
+    <a
+      href="https://scalar.com/download"
+      target="_blank">
+      Download it for free
+    </a>
+  </div>
+</template>
