+++ packages/api-client/src/components/ImportCollection/LoadingScreen.vue
@@ -1,0 +1,6 @@
+<template>
+  <div
+    class="w-full p-4 border text-center h-12 flex justify-center items-center bg-b-2">
+    <div>fetching…</div>
+  </div>
+</template>
