+++ packages/api-client/src/components/ImportCollection/OpenApiDocumentPreview.vue
@@ -1,0 +1,39 @@
+<script lang="ts" setup>
+import { getOpenApiDocumentVersion } from '@/components/ImportCollection/utils/getOpenApiDocumentVersion'
+import { ScalarCodeBlock, ScalarIcon } from '@scalar/components'
+import { isJsonString } from '@scalar/oas-utils/helpers'
+
+defineProps<{
+  content?: string | null
+}>()
+</script>
+<template>
+  <div
+    v-if="content"
+    class="flex gap-2 flex-col w-full">
+    <div class="flex gap-2 items-center">
+      <template v-if="getOpenApiDocumentVersion(content)">
+        <ScalarIcon
+          class="text-green"
+          icon="Checkmark"
+          size="sm" />
+        <div class="text-sm">
+          {{ getOpenApiDocumentVersion(content) }}
+        </div>
+      </template>
+      <template v-else>
+        <ScalarIcon
+          class="text-red"
+          icon="Error"
+          size="sm" />
+        <div>Oh, this doesn’t seem to be an OpenAPI/Swagger document:</div>
+      </template>
+    </div>
+    <div class="h-32 overflow-hidden border rounded">
+      <ScalarCodeBlock
+        class="bg-b-2"
+        :content="content ?? ''"
+        :lang="isJsonString(content) ? 'json' : 'yaml'" />
+    </div>
+  </div>
+</template>
