+++ packages/api-client/src/components/ImportCollection/utils/isUrl.ts
@@ -1,0 +1,4 @@
+/** Checks whether the given string is an URL */
+export function isUrl(input: string | null) {
+  return input && (input.startsWith('http://') || input.startsWith('https://'))
+}
